python/requirements.txt
flask < 2.0; python_version < "3.0"
markupsafe < 2.0; python_version < "3.0"
gunicorn < 20.0; python_version < "3.0"
opencensus-ext-flask; python_version < "3.0"
requests < 2.28.0; python_version < "3.0"
certifi < 2020.4.5.2; python_version < "3.0"
packaging < 21.0; python_version < "3.0"
pyparsing < 3.0.0; python_version < "3.0"
cachetools < 4.0; python_version < "3.0"
google-cloud-pubsub < 2.0; python_version < "3.0"
grpcio < 1.40; python_version < "3.0"
google-api-core < 2.0; python_version < "3.0"
opencensus-ext-jaeger; python_version < "3.0"
thrift > 0.10; python_version < "3.0"
